+++ quantify_scheduler/backends/qblox/instrument_compilers.py
@@ -139,9 +139,18 @@ class QcmModule(compiler_abc.QbloxBasebandModule):
         max_sequencers=NUMBER_OF_SEQUENCERS_QCM,
         max_awg_output_voltage=2.5,
         mixer_dc_offset_range=BoundedParameter(min_val=-2.5, max_val=2.5, units="V"),
-        valid_ios=[f"complex_output_{i}" for i in range(2)]
-        + [f"real_output_{i}" for i in range(4)]
-        + [f"digital_output_{i}" for i in range(4)],
+        io_name_to_connected_io_indices={
+            "complex_output_0": (0, 1),
+            "complex_output_1": (2, 3),
+            "real_output_0": (0,),
+            "real_output_1": (1,),
+            "real_output_2": (2,),
+            "real_output_3": (3,),
+            "digital_output_0": (0,),
+            "digital_output_1": (1,),
+            "digital_output_2": (2,),
+            "digital_output_3": (3,),
+        },
     )
 
 
@@ -155,11 +164,18 @@ class QrmModule(compiler_abc.QbloxBasebandModule):
         max_sequencers=NUMBER_OF_SEQUENCERS_QRM,
         max_awg_output_voltage=0.5,
         mixer_dc_offset_range=BoundedParameter(min_val=-0.5, max_val=0.5, units="V"),
-        valid_ios=[f"complex_output_{i}" for i in [0]]
-        + [f"real_output_{i}" for i in range(2)]
-        + [f"complex_input_{i}" for i in [0]]
-        + [f"real_input_{i}" for i in range(2)]
-        + [f"digital_output_{i}" for i in range(4)],
+        io_name_to_connected_io_indices={
+            "complex_output_0": (0, 1),
+            "complex_input_0": (0, 1),
+            "real_output_0": (0,),
+            "real_output_1": (1,),
+            "real_input_0": (0,),
+            "real_input_1": (1,),
+            "digital_output_0": (0,),
+            "digital_output_1": (1,),
+            "digital_output_2": (2,),
+            "digital_output_3": (3,),
+        },
     )
 
 
@@ -172,9 +188,13 @@ class QcmRfModule(compiler_abc.QbloxRFModule):
         max_sequencers=NUMBER_OF_SEQUENCERS_QCM,
         max_awg_output_voltage=None,
         mixer_dc_offset_range=BoundedParameter(min_val=-50, max_val=50, units="mV"),
-        valid_ios=[f"complex_output_{i}" for i in range(2)]
-        + [f"digital_output_{i}" for i in range(2)],
-        output_map={
+        io_name_to_connected_io_indices={
+            "complex_output_0": (0, 1),
+            "complex_output_1": (2, 3),
+            "digital_output_0": (0,),
+            "digital_output_1": (1,),
+        },
+        io_name_to_digital_marker={
             "complex_output_0": 0b0001,
             "complex_output_1": 0b0010,
         },
@@ -191,9 +211,12 @@ class QrmRfModule(compiler_abc.QbloxRFModule):
         max_sequencers=NUMBER_OF_SEQUENCERS_QRM,
         max_awg_output_voltage=None,
         mixer_dc_offset_range=BoundedParameter(min_val=-50, max_val=50, units="mV"),
-        valid_ios=[f"complex_output_{i}" for i in [0]]
-        + [f"digital_output_{i}" for i in range(2)]
-        + [f"complex_input_{i}" for i in [0]],
+        io_name_to_connected_io_indices={
+            "complex_output_0": (0, 1),
+            "complex_input_0": (0, 1),
+            "digital_output_0": (0,),
+            "digital_output_1": (1,),
+        },
         default_marker=0b0011,
     )