+++ quantify_scheduler/backends/qblox/operation_handling/pulses.py
@@ -10,6 +10,7 @@ from typing import Any, Dict, Optional
 import numpy as np
 
 from quantify_scheduler.backends.qblox import constants, helpers, q1asm_instructions
+from quantify_scheduler.backends.qblox.enums import IoMode
 from quantify_scheduler.backends.qblox.operation_handling.base import IOperationStrategy
 from quantify_scheduler.backends.qblox.qasm_program import QASMProgram
 from quantify_scheduler.backends.types import qblox as types
@@ -27,11 +28,11 @@ class PulseStrategyPartial(IOperationStrategy):
     operation_info
         The operation info that corresponds to this pulse.
     io_mode
-        Either "real", "imag" or complex depending on whether the signal affects
-        only path0, path1 or both.
+        Either :attr:`.IoMode.REAL`, :attr:`.IoMode.IMAG` or :attr:`.IoMode.COMPLEX`
+        depending on whether the signal affects only path0, path1 or both.
     """
 
-    def __init__(self, operation_info: types.OpInfo, io_mode: str):
+    def __init__(self, operation_info: types.OpInfo, io_mode: IoMode):
         self._pulse_info: types.OpInfo = operation_info
         self.io_mode = io_mode
 
@@ -59,12 +60,15 @@ class GenericPulseStrategy(PulseStrategyPartial):
     operation_info
         The operation info that corresponds to this pulse.
     io_mode
-        Either "real", "imag" or "complex" depending on whether the signal affects
-        only path0, path1 or both, respectively.
+        Either :attr:`.IoMode.REAL`, :attr:`.IoMode.IMAG` or :attr:`.IoMode.COMPLEX`
+        depending on whether the signal affects only path0, path1 or both.
     """
 
-    def __init__(self, operation_info: types.OpInfo, io_mode: str):
-        super().__init__(operation_info, io_mode)
+    def __init__(self, operation_info: types.OpInfo, io_mode: IoMode):
+        super().__init__(
+            operation_info=operation_info,
+            io_mode=io_mode,
+        )
 
         self._amplitude_path0: Optional[float] = None
         self._amplitude_path1: Optional[float] = None
@@ -78,8 +82,9 @@ class GenericPulseStrategy(PulseStrategyPartial):
         """
         Generates the data and adds them to the ``wf_dict`` (if not already present).
 
-        In complex mode, real-valued data is produced on sequencer path0 (:math:`I_\\text{IF}`)
-        and imaginary data on sequencer path1 (:math:`Q_\\text{IF}`) after the NCO mixing.
+        In complex mode (``io_mode == IoMode.COMPLEX``), the NCO produces real-valued data
+        (:math:`I_\\text{IF}`) on sequencer path0 and imaginary data (:math:`Q_\\text{IF}`)
+        on sequencer path1.
 
         .. math::
             \\underbrace{\\begin{bmatrix}
@@ -88,20 +93,44 @@ class GenericPulseStrategy(PulseStrategyPartial):
             \\begin{bmatrix}
             I \\\\
             Q \\end{bmatrix} =
+            \\begin{bmatrix}
+            I \\cdot \\cos\\omega t - Q \\cdot\\sin\\omega t \\\\
+            I \\cdot \\sin\\omega t + Q \\cdot\\cos\\omega t \\end{bmatrix}
             \\begin{matrix}
-            \\overbrace{ I \\cdot \\cos\\omega t - Q \\cdot\\sin\\omega t}^{\\small \\textbf{real} \\Rightarrow \\text{path0}} \\\\
-            \\underbrace{I \\cdot \\sin\\omega t + Q \\cdot\\cos\\omega t}_{\\small \\textbf{imag} \\Rightarrow \\text{path1}} \\end{matrix} =
+            \\ \\text{(path0)} \\\\
+            \\ \\text{(path1)} \\end{matrix}
+            =
             \\begin{bmatrix}
             I_\\text{IF} \\\\
             Q_\\text{IF} \\end{bmatrix}
 
-        In real mode, :math:`I_\\text{IF}` can be produced on either
-        path0 (``io_mode == "real"``) or path1 (``io_mode == "imag"``).
 
-        For ``io_mode == imag``, the real-valued input (:math:`I`) on path0 is
-        swapped with imaginary input (:math:`Q`) on path1. We multiply :math:`Q` by -1
-        (via ``amp_imag``) to undo the 90-degree phase shift resulting from swapping the
-        NCO input paths.
+        In real mode (``io_mode == IoMode.REAL``), the NCO produces :math:`I_\\text{IF}` on
+        path0 
+
+
+        .. math::
+            \\underbrace{\\begin{bmatrix}
+            \\cos\\omega t & -\\sin\\omega t \\\\
+            \\sin\\omega t & \\phantom{-}\\cos\\omega t \\end{bmatrix}}_\\text{NCO}
+            \\begin{bmatrix}
+            I \\\\
+            Q \\end{bmatrix}  =
+            \\begin{bmatrix}
+            I \\cdot \\cos\\omega t - Q \\cdot\\sin\\omega t\\\\
+             - \\end{bmatrix}
+            \\begin{matrix}
+            \\ \\text{(path0)} \\\\
+            \\ \\text{(path1)} \\end{matrix}
+            =
+            \\begin{bmatrix}
+            I_\\text{IF} \\\\
+            - \\end{bmatrix}
+        
+        
+        while in imaginary mode (``io_mode == IoMode.IMAG``), the NCO produces 
+        :math:`I_\\text{IF}` on path1.
+
 
         .. math::
             \\underbrace{\\begin{bmatrix}
@@ -110,12 +139,19 @@ class GenericPulseStrategy(PulseStrategyPartial):
             \\begin{bmatrix}
             -Q \\\\
             I \\end{bmatrix}  =
+            \\begin{bmatrix}
+             - \\\\
+            -Q \\cdot \\sin\\omega t + I \\cdot\\cos\\omega t \\end{bmatrix}
             \\begin{matrix}
-            \\\\
-            \\underbrace{-Q \\cdot \\sin\\omega t + I \\cdot\\cos\\omega t}_{\\small \\textbf{real} \\Rightarrow \\text{path1}} \\end{matrix}=
+            \\ \\text{(path0)} \\\\
+            \\ \\text{(path1)} \\end{matrix} =
             \\begin{bmatrix}
             - \\\\
             I_\\text{IF} \\end{bmatrix}
+
+        Note that the fields marked with `-` represent waveforms that are not relevant
+        for the mode.
+
 
         Parameters
         ----------
@@ -127,7 +163,7 @@ class GenericPulseStrategy(PulseStrategyPartial):
         ------
         ValueError
             Data is complex (has an imaginary component), but the io_mode is not set
-            to "complex".
+            to :attr:`.IoMode.COMPLEX`.
         """  # pylint: disable=line-too-long  # noqa: D301
         op_info = self.operation_info
         waveform_data = helpers.generate_waveform_data(
@@ -136,7 +172,7 @@ class GenericPulseStrategy(PulseStrategyPartial):
         waveform_data, amp_real, amp_imag = normalize_waveform_data(waveform_data)
         self._waveform_len = len(waveform_data)
 
-        if np.any(np.iscomplex(waveform_data)) and not self.io_mode == "complex":
+        if np.any(np.iscomplex(waveform_data)) and not self.io_mode == IoMode.COMPLEX:
             raise ValueError(
                 f"Complex valued {str(op_info)} detected but the sequencer"
                 f" is not expecting complex input. This can be caused by "
@@ -161,8 +197,11 @@ class GenericPulseStrategy(PulseStrategyPartial):
             else None
         )
 
-        # Update self._waveform_index and self._amplitude_path
-        if self.io_mode == "imag":
+        # Imaginary mode requires swapping the real- and imaginary-valued NCOs
+        # inputs, and this produces a 90-degree phase shift that we undo by
+        # multiplying Q by -1 (note this is done for correctness/completeness,
+        # path0 is not used in imaginary mode).
+        if self.io_mode == IoMode.IMAG:
             self._waveform_index0, self._waveform_index1 = idx_imag, idx_real
             self._amplitude_path0, self._amplitude_path1 = (
                 -amp_imag,  # Multiply by -1 to undo 90-degree shift
@@ -224,9 +263,15 @@ class MarkerPulseStrategy(PulseStrategyPartial):
         qasm_program
             The QASMProgram to add the assembly instructions to.
         """
-        if self.io_mode != "digital":
+        if self.io_mode != IoMode.DIGITAL:
+            port = self.operation_info.data.get("port")
+            clock = self.operation_info.data.get("clock")
+
             raise ValueError(
-                f"MarkerPulseStrategy can only be used with digital IO, not {self.io_mode}. "
+                f"{MarkerPulseStrategy.__name__} can only be used with "
+                f"digital IO, not io_mode '{self.io_mode}'. Please make sure that "
+                f"'digital' keyword is included in the io_name in the hardware configuration "
+                f"for port-clock combination '{port}-{clock}'."
                 f"Operation causing exception: {self.operation_info}"
             )
         duration = round(self.operation_info.duration * 1e9)