+++ tests/scheduler/backends/qblox/operation_handling/test_pulses.py
@@ -10,20 +10,21 @@
 # pylint: disable=too-many-locals
 # pylint: disable=too-many-arguments
 
-import pytest
-import numpy as np
 import re
 
-from quantify_scheduler import waveforms, Schedule
+import numpy as np
+import pytest
+
+from quantify_scheduler import Schedule, waveforms
 from quantify_scheduler.backends import SerialCompiler
-from quantify_scheduler.helpers.waveforms import normalize_waveform_data
-from quantify_scheduler.backends.types import qblox as types
 from quantify_scheduler.backends.qblox import constants
+from quantify_scheduler.backends.qblox.enums import IoMode
 from quantify_scheduler.backends.qblox.operation_handling import pulses
+from quantify_scheduler.backends.types import qblox as types
+from quantify_scheduler.helpers.waveforms import normalize_waveform_data
 from quantify_scheduler.operations.gate_library import Measure
 from quantify_scheduler.operations.pulse_library import MarkerPulse, SquarePulse
 from quantify_scheduler.resources import ClockResource
-
 from tests.scheduler.instrument_coordinator.components.test_qblox import (  # pylint: disable=unused-import
     make_cluster_component,
 )
@@ -36,19 +37,23 @@ from .empty_qasm_program import (  # pylint: disable=unused-import
 class TestGenericPulseStrategy:
     def test_constructor(self):
         pulses.GenericPulseStrategy(
-            types.OpInfo(name="", data={}, timing=0), io_mode="real"
+            operation_info=types.OpInfo(name="", data={}, timing=0),
+            io_mode=IoMode.REAL,
         )
 
     def test_operation_info_property(self):
         # arrange
-        op_info = types.OpInfo(name="", data={}, timing=0)
-        strategy = pulses.GenericPulseStrategy(op_info, io_mode="real")
+        operation_info = types.OpInfo(name="", data={}, timing=0)
+        strategy = pulses.GenericPulseStrategy(
+            operation_info=operation_info,
+            io_mode=IoMode.REAL,
+        )
 
         # act
         from_property = strategy.operation_info
 
         # assert
-        assert op_info == from_property
+        assert operation_info == from_property
 
     @pytest.mark.parametrize(
         "wf_func, wf_func_path, wf_kwargs",
@@ -75,8 +80,11 @@ class TestGenericPulseStrategy:
         duration = 24e-9
         data = {"wf_func": wf_func_path, "duration": duration, **wf_kwargs}
 
-        op_info = types.OpInfo(name="", data=data, timing=0)
-        strategy = pulses.GenericPulseStrategy(op_info, io_mode="real")
+        operation_info = types.OpInfo(name="", data=data, timing=0)
+        strategy = pulses.GenericPulseStrategy(
+            operation_info=operation_info,
+            io_mode=IoMode.REAL,
+        )
         wf_dict = {}
         t_test = np.linspace(0, duration, int(duration * constants.SAMPLING_RATE))
 
@@ -107,8 +115,10 @@ class TestGenericPulseStrategy:
             "phase": 0,
         }
 
-        op_info = types.OpInfo(name="", data=data, timing=0)
-        strategy = pulses.GenericPulseStrategy(op_info, io_mode="complex")
+        strategy = pulses.GenericPulseStrategy(
+            operation_info=types.OpInfo(name="", data=data, timing=0),
+            io_mode=IoMode.COMPLEX,
+        )
         wf_dict = {}
         t_test = (
             np.arange(0, int(round(duration * constants.SAMPLING_RATE)), 1)
@@ -160,8 +170,10 @@ class TestGenericPulseStrategy:
         duration = 24e-9
         data = {"wf_func": wf_func_path, "duration": duration, **wf_kwargs}
 
-        op_info = types.OpInfo(name="", data=data, timing=0)
-        strategy = pulses.GenericPulseStrategy(op_info, io_mode="imag")
+        strategy = pulses.GenericPulseStrategy(
+            operation_info=types.OpInfo(name="", data=data, timing=0),
+            io_mode=IoMode.IMAG,
+        )
         wf_dict = {}
         t_test = np.arange(0, duration, step=1e-9)
 
@@ -182,7 +194,7 @@ class TestGenericPulseStrategy:
 
     @pytest.mark.parametrize(
         "io_mode",
-        ["real", "imag"],
+        [IoMode.REAL, IoMode.IMAG],
     )
     def test_exception_wrong_mode(self, io_mode):
         # arrange
@@ -196,8 +208,10 @@ class TestGenericPulseStrategy:
             "phase": 0,
         }
 
-        op_info = types.OpInfo(name="test_pulse_name", data=data, timing=0)
-        strategy = pulses.GenericPulseStrategy(op_info, io_mode=io_mode)
+        strategy = pulses.GenericPulseStrategy(
+            operation_info=types.OpInfo(name="test_pulse_name", data=data, timing=0),
+            io_mode=io_mode,
+        )
         wf_dict = {}
 
         # act
@@ -218,7 +232,6 @@ class TestGenericPulseStrategy:
 
     def test_insert_qasm(self, empty_qasm_program_qcm):
         # arrange
-        qasm = empty_qasm_program_qcm
         duration = 24e-9
         wf_func_path = "quantify_scheduler.waveforms.drag"
         wf_kwargs = {
@@ -230,41 +243,78 @@ class TestGenericPulseStrategy:
         }
         data = {"wf_func": wf_func_path, "duration": duration, **wf_kwargs}
 
-        op_info = types.OpInfo(name="test_pulse", data=data, timing=0)
-        strategy = pulses.GenericPulseStrategy(op_info, io_mode="complex")
+        strategy = pulses.GenericPulseStrategy(
+            operation_info=types.OpInfo(name="test_pulse", data=data, timing=0),
+            io_mode=IoMode.COMPLEX,
+        )
         strategy.generate_data(wf_dict={})
 
         # act
-        strategy.insert_qasm(qasm)
+        strategy.insert_qasm(empty_qasm_program_qcm)
 
         # assert
         line0 = ["", "set_awg_gain", "32212,20355", "# setting gain for test_pulse"]
         line1 = ["", "play", "0,1,4", "# play test_pulse (24 ns)"]
-        assert qasm.instructions[0] == line0
-        assert qasm.instructions[1] == line1
+        assert empty_qasm_program_qcm.instructions[0] == line0
+        assert empty_qasm_program_qcm.instructions[1] == line1
 
 
 class TestMarkerPulseStrategy:
     def test_constructor(self):
         pulses.MarkerPulseStrategy(
-            types.OpInfo(name="", data={}, timing=0), io_mode="real"
+            operation_info=types.OpInfo(name="", data={}, timing=0),
+            io_mode=IoMode.DIGITAL,
         )
+
+    def test_insert_qasm_exception(self, empty_qasm_program_qcm):
+        duration = 24e-9
+        wf_func_path = "quantify_scheduler.waveforms.drag"
+        wf_kwargs = {
+            "G_amp": 1.0,
+            "D_amp": 1.0,
+            "duration": 24e-9,
+            "nr_sigma": 3,
+            "phase": 0,
+        }
+        data = {"wf_func": wf_func_path, "duration": duration, **wf_kwargs}
+
+        strategy = pulses.MarkerPulseStrategy(
+            operation_info=types.OpInfo(name="test_pulse", data=data, timing=0),
+            io_mode=IoMode.COMPLEX,
+        )
+        strategy.generate_data(wf_dict={})
+
+        with pytest.raises(
+            ValueError,
+            match=re.escape(
+                "MarkerPulseStrategy can only be used with digital IO, "
+                "not io_mode 'complex'. Please make sure that 'digital' keyword is included "
+                "in the io_name in the hardware configuration for port-clock combination"
+                " 'None-None'.Operation causing exception: Pulse \"test_pulse\" (t0=0, duration=2.4e-08)"
+            ),
+        ):
+            strategy.insert_qasm(empty_qasm_program_qcm)
 
     def test_operation_info_property(self):
         # arrange
-        op_info = types.OpInfo(name="", data={}, timing=0)
-        strategy = pulses.MarkerPulseStrategy(op_info, io_mode="real")
+        operation_info = types.OpInfo(name="", data={}, timing=0)
+        strategy = pulses.MarkerPulseStrategy(
+            operation_info=operation_info,
+            io_mode=IoMode.DIGITAL,
+        )
 
         # act
         from_property = strategy.operation_info
 
         # assert
-        assert op_info == from_property
+        assert operation_info == from_property
 
     def test_generate_data(self):
         # arrange
-        op_info = types.OpInfo(name="", data={}, timing=0)
-        strategy = pulses.MarkerPulseStrategy(op_info, io_mode="real")
+        strategy = pulses.MarkerPulseStrategy(
+            operation_info=types.OpInfo(name="", data={}, timing=0),
+            io_mode=IoMode.DIGITAL,
+        )
 
         # act
         # pylint: disable=assignment-from-none